+++ src/cambium/core.clj
@@ -9,7 +9,6 @@
 
 (ns cambium.core
   (:require
-    [clojure.edn                :as edn]
     [clojure.tools.logging      :as ctl]
     [clojure.tools.logging.impl :as ctl-impl]
     [cambium.internal           :as i]
@@ -28,58 +27,13 @@
 
 
 (def ^:redef stringify-val
-  "Arity-1 fn to convert MDC value into a string. By default this carries out a plain string conversion.
-  See: encode-val"
+  "Arity-1 fn to convert MDC value into a string. By default this carries out a plain string conversion."
   i/as-str)
 
 
 (def ^:redef destringify-val
-  "Arity-1 fn to convert MDC string back to original value. By default this simply returns the stored String value.
-  See: decode-val"
+  "Arity-1 fn to convert MDC string back to original value. By default this simply returns the stored String value."
   identity)
-
-
-;; ----- EDN based codec helper -----
-
-
-(defn encode-val
-  "Encode MDC value as string such that it retains type information.
-  See: decode-val"
-  (^String [object-encoder v]
-    (let [hint-str (fn ^String [^String hint v]
-                     (let [^StringBuilder sb (StringBuilder. 15)]
-                       (.append sb hint)
-                       (.append sb v)
-                       (.toString sb)))]
-      (cond
-        (string? v)  v  ; do not follow escape-safety due to performance
-        (instance?
-          clojure.lang.Named v) (name v)
-        (integer? v) (hint-str "^long "    v)
-        (float? v)   (hint-str "^double "  v)
-        (instance?
-          Boolean v) (hint-str "^boolean " v)
-        :otherwise   (hint-str "^object "  (object-encoder v)))))
-  (^String [v]
-    (encode-val pr-str v)))
-
-
-(defn decode-val
-  "Decode MDC string value into the correct original type.
-  See: encode-val"
-  ([object-decoder ^String s]
-    (cond
-      (nil? s)             s
-      (= 0 (.length s))    s
-      (= \^ (.charAt s 0)) (cond
-                             (.startsWith s "^long ")    (try (Long/parseLong     (subs s 6)) (catch Exception e -314))
-                             (.startsWith s "^double ")  (try (Double/parseDouble (subs s 8)) (catch Exception e -3.14))
-                             (.startsWith s "^boolean ") (Boolean/parseBoolean    (subs s 9))
-                             (.startsWith s "^object ")  (try (object-decoder (subs s 8)) (catch Exception e (str e)))
-                             :otherwise                  s)
-      :otherwise           s))
-  ([^String s]
-    (decode-val edn/read-string s)))
 
 
 ;; ----- MDC handling -----
@@ -95,8 +49,7 @@
 
 (defn ^:redef context-val
   "Return the value of the specified key from the current context; behavior for non-existent keys would be
-  implementation dependent - it may return nil or may throw exception.
-  See: cambium.core/encode-val, cambium.core/decode-val"
+  implementation dependent - it may return nil or may throw exception."
   [k]
   (destringify-val (MDC/get (stringify-key k))))