+++ src/cambium/nested.clj
@@ -8,10 +8,13 @@
 
 
 (ns cambium.nested
-  "API for handling nested context.
-  Vars 'cambium.core/stringify-val' and 'cambium.core/destringify-val' must be redefined to preserve structure.
-  See: cambium.core/encode-val, cambium.core/decode-val"
+  "Utility fns for handling nested context. Mostly useful for redefining vars in 'cambium.core' namespace:
+  - cambium.core/stringify-val
+  - cambium.core/destringify-val
+  - cambium.core/context-val
+  - cambium.core/merge-logging-context!"
   (:require
+    [clojure.edn                :as edn]
     [clojure.tools.logging      :as ctl]
     [clojure.tools.logging.impl :as ctl-impl]
     [cambium.core     :as c]
@@ -20,6 +23,52 @@
   (:import
     [java.util HashMap Map$Entry]
     [org.slf4j MDC]))
+
+
+;; ----- Codec (default: EDN codec) helper -----
+
+
+(defn encode-val
+  "Encode MDC value as string such that it retains type information. May be used to redefine cambium.core/stringify-val.
+  See: decode-val"
+  (^String [object-encoder v]
+    (let [hint-str (fn ^String [^String hint v]
+                     (let [^StringBuilder sb (StringBuilder. 15)]
+                       (.append sb hint)
+                       (.append sb v)
+                       (.toString sb)))]
+      (cond
+        (string? v)  v  ; do not follow escape-safety due to performance
+        (instance?
+          clojure.lang.Named v) (name v)
+        (integer? v) (hint-str "^long "    v)
+        (float? v)   (hint-str "^double "  v)
+        (instance?
+          Boolean v) (hint-str "^boolean " v)
+        :otherwise   (hint-str "^object "  (object-encoder v)))))
+  (^String [v]
+    (encode-val pr-str v)))
+
+
+(defn decode-val
+  "Decode MDC string value into the correct original type. May be used to redefine cambium.core/destringify-val.
+  See: encode-val"
+  ([object-decoder ^String s]
+    (cond
+      (nil? s)             s
+      (= 0 (.length s))    s
+      (= \^ (.charAt s 0)) (cond
+                             (.startsWith s "^long ")    (try (Long/parseLong     (subs s 6)) (catch Exception e -314))
+                             (.startsWith s "^double ")  (try (Double/parseDouble (subs s 8)) (catch Exception e -3.14))
+                             (.startsWith s "^boolean ") (Boolean/parseBoolean    (subs s 9))
+                             (.startsWith s "^object ")  (try (object-decoder (subs s 8)) (catch Exception e (str e)))
+                             :otherwise                  s)
+      :otherwise           s))
+  ([^String s]
+    (decode-val edn/read-string s)))
+
+
+;; ----- Nested context navigation -----
 
 
 (defn nested-context-val