+++ test/cambium/core_test.clj
@@ -27,20 +27,7 @@
   (testing "custom loggers"
     (tu/metrics {:latency-ns 430 :module "registration"} "op.latency")
     (tu/metrics {:module "registration"} (ex-info "some error" {:data :foo}) "internal error")
-    (tu/txn-metrics {:module "order-fetch"} "Fetched order #4568"))
-  (testing "type-safe encoding"
-    (let [sk c/stringify-key]
-      (with-redefs [c/stringify-key (fn ^String [x] (.replace ^String (sk x) \- \_))
-                    c/stringify-val c/encode-val
-                    c/destringify-val c/decode-val]
-        (c/info "hello")
-        (c/info {:foo-k "bar" :baz 10 :qux true} "hello with context")
-        (c/with-logging-context {:extra-k "context" "some-data" [1 2 :three 'four]}
-          (is (= (c/get-context) {"extra_k" "context" "some_data" [1 2 :three 'four]}))
-          (is (= (c/context-val :extra-k) "context"))
-          (is (nil? (c/context-val "foo")))
-          (c/info {:foo "bar"} "hello with wrapped context"))
-        (c/error {} (ex-info "some error" {:data :foo}) "internal error")))))
+    (tu/txn-metrics {:module "order-fetch"} "Fetched order #4568")))
 
 
 (deftest test-context-propagation
@@ -75,16 +62,3 @@
       ((c/wrap-logging-context context-old f) :dummy :arg)
       ((comp (partial c/wrap-logging-context context-new) (c/wrap-logging-context context-old f)))
       (is (nil? (c/context-val :foo))))))
-
-
-(deftest test-codec
-  (let [payload (c/encode-val :foo)]  (is (= "foo" payload))           (is (= "foo" (c/decode-val payload))))
-  (let [payload (c/encode-val 'foo)]  (is (= "foo" payload))           (is (= "foo" (c/decode-val payload))))
-  (let [payload (c/encode-val "foo")] (is (= "foo" payload))           (is (= "foo" (c/decode-val payload))))
-  (let [payload (c/encode-val 10)]    (is (= "^long 10" payload))      (is (= 10    (c/decode-val payload))))
-  (let [payload (c/encode-val 1.2)]   (is (= "^double 1.2" payload))   (is (= 1.2   (c/decode-val payload))))
-  (let [payload (c/encode-val true)]  (is (= "^boolean true" payload)) (is (= true  (c/decode-val payload))))
-  (let [payload (c/encode-val nil)]   (is (= "^object nil" payload))   (is (= nil   (c/decode-val payload))))
-  (let [payload (c/encode-val
-                  [1 :two 'four])]    (is (= "^object [1 :two four]"
-                                            payload))                  (is (= [1 :two 'four] (c/decode-val payload)))))